+++ flake.nix
@@ -3,9 +3,9 @@
 
   inputs = {
     nixpkgs.url = "github:nixos/nixpkgs/nixpkgs-22.11-darwin";
-    cardano-wallet-unpatched.url = "github:cardano-foundation/cardano-wallet/v2024-05-05";
+    cardano-wallet-unpatched.url = "github:cardano-foundation/cardano-wallet/v2024-07-27";
     cardano-wallet-unpatched.flake = false; # otherwise, +10k quadratic dependencies in flake.lock…
-    cardano-playground.url = "github:input-output-hk/cardano-playground";
+    cardano-playground.url = "github:input-output-hk/cardano-playground/next-2024-07-24";
     cardano-playground.flake = false; # otherwise, +9k dependencies in flake.lock…
     cardano-shell.url = "github:input-output-hk/cardano-shell/0d1d5f036c73d18e641412d2c58d4acda592d493";
     cardano-shell.flake = false;
@@ -13,6 +13,8 @@
     tullia.flake = false; # otherwie, +1k dependencies in flake.lock…
     flake-compat.url = "github:input-output-hk/flake-compat";
     flake-compat.flake = false;
+    nix-bundle-exe.url = "github:3noch/nix-bundle-exe";
+    nix-bundle-exe.flake = false;
   };
 
   outputs = inputs: let