+++ nix/devshells.nix
@@ -47,8 +47,9 @@ let
       darwin.apple_sdk.frameworks.AppKit
       darwin.cctools
       xcbuild
+      perl
     ] else [
-      internal.electronBin
+      internal.relocatableElectron
       winePackages.minimal
     ]));
     name = "daedalus";