+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## vNext
 
+### Chores
+
+- Updated `@cardano-foundation/ledgerjs-hw-app-cardano` to version `7.1.3` and `@trezor/connect` to version `9.3.0` ([PR 3093](https://github.com/input-output-hk/daedalus/pull/3215))
+
 ## 5.4.0
 
 ### Fixes